[rocprofiler-compute] Extract metric unit tests into tests/test_metrics.py
Move the metric-focused unit tests out of tests/test_analyze_commands.py
into a dedicated tests/test_metrics.py and group them into per-module
classes (TestAggregation, TestExpression, TestEvaluationPipeline,
TestMetricEvaluator). Tests are moved verbatim - no behavior change -
and the new file is registered in CMakeLists.txt and the coverage test
list. Per-test sys.path insertion is dropped because conftest.py
already exposes src on sys.path.

Made-with: Cursor
