+++ projects/rocprofiler-compute/CMakeLists.txt
@@ -539,6 +539,18 @@ add_test(
 )
 
 # -----------------------------------
+# Metrics unit tests
+# -----------------------------------
+
+add_test(
+    NAME test_metrics
+    COMMAND
+        ${PYTHON_TEST_COMMAND} -m pytest ${STANDALONEBINARY_TEST_OPTION}
+        --junitxml=tests/test_metrics.xml ${COV_OPTION} tests/test_metrics.py
+        ${WORKING_DIR_OPTION}
+)
+
+# -----------------------------------
 # Autogenerated configuration tests
 # -----------------------------------
 
@@ -655,6 +667,7 @@ if(${ENABLE_COVERAGE})
         test_pc_sampling
         test_analyze_commands
         test_analyze_workloads
+        test_metrics
         test_L1_cache_counters
         test_num_xcds_spec_class
         test_num_xcds_cli_output
